+++ domain/src/main/java/com/danbam/domain/entity/MoviePeopleDetailEntity.kt
@@ -1,7 +1,7 @@
 package com.danbam.domain.entity
 
 data class MoviePeopleDetailEntity(
-    val idx: Int,
+    val idx: Long,
     val name: String,
     val profileUrl: String,
     val movieList: List<MovieEntity>
