+++ domain/src/main/java/com/danbam/domain/usecase/movie/MoviePeopleDetailUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class MoviePeopleDetailUseCase @Inject constructor(
     private val movieRepository: MovieRepository
 ) {
-    suspend operator fun invoke(actorType: String, idx: Int) = runCatching {
+    suspend operator fun invoke(actorType: String, idx: Long) = runCatching {
         movieRepository.moviePeopleDetail(actorType = actorType, idx = idx)
     }
 }
